requirements.txt
environs
django
djangorestframework
psycopg2-binary
gunicorn